implement iterator for StructArray
