+++ src/array/struct_/mod.rs
@@ -3,11 +3,12 @@ use std::sync::Arc;
 use crate::{
     bitmap::Bitmap,
     datatypes::{DataType, Field},
-    error::Result,
-    ffi,
 };
 
-use super::{ffi::ToFfi, new_empty_array, new_null_array, Array, FromFfi};
+use super::{new_empty_array, new_null_array, Array};
+
+mod ffi;
+mod iterator;
 
 /// A [`StructArray`] is a nested [`Array`] with an optional validity representing
 /// multiple [`Array`] with the same number of rows.
@@ -220,45 +221,5 @@ impl std::fmt::Display for StructArray {
             writeln!(f, "{}: {},", field.name(), column)?;
         }
         write!(f, "}}")
-    }
-}
-
-unsafe impl ToFfi for StructArray {
-    fn buffers(&self) -> Vec<Option<std::ptr::NonNull<u8>>> {
-        vec![self.validity.as_ref().map(|x| x.as_ptr())]
-    }
-
-    fn children(&self) -> Vec<Arc<dyn Array>> {
-        self.values.clone()
-    }
-
-    fn offset(&self) -> Option<usize> {
-        Some(
-            self.validity
-                .as_ref()
-                .map(|bitmap| bitmap.offset())
-                .unwrap_or_default(),
-        )
-    }
-
-    fn to_ffi_aligned(&self) -> Self {
-        self.clone()
-    }
-}
-
-impl<A: ffi::ArrowArrayRef> FromFfi<A> for StructArray {
-    unsafe fn try_from_ffi(array: A) -> Result<Self> {
-        let data_type = array.field().data_type().clone();
-        let fields = Self::get_fields(&data_type);
-
-        let validity = unsafe { array.validity() }?;
-        let values = (0..fields.len())
-            .map(|index| {
-                let child = array.child(index)?;
-                Ok(ffi::try_from(child)?.into())
-            })
-            .collect::<Result<Vec<Arc<dyn Array>>>>()?;
-
-        Ok(Self::from_data(data_type, values, validity))
     }
 }
